+++ CHANGELOG.md
@@ -8,3 +8,4 @@ All notable changes to this project will be documented in this file.
 
 - Horizontal scene navigation option that displays scenes as pills across the top of the screen instead of the v13 vertical dropdown
 - Collapsible hotbar that slides down when you want it out of the way
+- Collapsible chat notifications that slide right to hide when you don't need them visible